Fix enchancePrisma leaking to browser on certain cases
