+++ packages/blitz-rpc/.changeset/mean-dogs-tease.md
@@ -1,0 +1,6 @@
+---
+"blitz": patch
+"@blitzjs/rpc": patch
+---
+
+Fix enhancePrisma is undefined when importing types in the browser
