+++ packages/blitz/src/index-browser.tsx
@@ -83,3 +83,4 @@ export * from "./ts-utils"
 export * from "./types"
 export * from "./errors"
 export * from "./zod-utils"
+export * from "./prisma-utils"
