+++ packages/blitz/src/index-server.ts
@@ -7,6 +7,7 @@ import {Ctx} from "./types"
 export * from "./index-browser"
 export * from "./types"
 export * from "./prisma-utils"
+export * from "./run-prisma"
 export * from "./middleware"
 export * from "./paginate"
 export {baseLogger, newLine, log} from "./logging"
