+++ packages/blitz/src/prisma-utils.ts
@@ -1,7 +1,5 @@
 import {spawn} from "cross-spawn"
 import which from "npm-which"
-import {Readable} from "stream"
-import pEvent from "p-event"
 
 export interface Constructor<T = unknown> {
   new (...args: never[]): T
@@ -53,27 +51,4 @@ export const enhancePrisma = <TPrismaClientCtor extends Constructor>(
       return globalThis._blitz_prismaClient
     },
   })
-}
-
-export const runPrisma = async (args: string[], silent = false) => {
-  const prismaBin = which(process.cwd()).sync("prisma")
-
-  const cp = spawn(prismaBin, args, {
-    stdio: silent ? "pipe" : "inherit",
-    env: process.env,
-  })
-
-  const cp_stderr: string[] = []
-  if (silent) {
-    cp?.stderr?.on("data", (chunk: Readable) => {
-      cp_stderr.push(chunk.toString())
-    })
-  }
-
-  const code = await pEvent(cp, "exit", {rejectionEvents: []})
-
-  return {
-    success: code === 0,
-    stderr: silent ? cp_stderr.join("") : undefined,
-  }
 }
